tests/wpt/mozilla/tests/mozilla/element-getcomputedstyle.html

<html>
  <head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      // Issue #7268 - getComputedStyle should work on non-HTML elements, too
      test(function() {
        var el = document.createElementNS("http://example.com", "a");
        document.body.appendChild(el);
        var style = window.getComputedStyle(el);

        assert_equals(style.opacity, "1");
      }, "getComputedStyle should work on non-HTML elements");
    </script>
  </body>
</html>
